Split editor grid widget handling
