+++ CMakeLists.txt
@@ -174,6 +174,7 @@ target_sources(
         src/game_data_editor_player_start.c
         src/game_data_editor_selection_queries.c
         src/game_data_editor_debug_primitives.c
+        src/game_data_editor_grid_widget.c
         src/game_data_grid_actions.c
         src/game_data_grid_runtime.c
         src/game_data_world_model_instances.c
