+++ src/game_data_editor_internal.h
@@ -39,6 +39,8 @@ bool editor_pick_selection_from_json(const slayer3d_game_data_runtime *runtime, 
                                      const slayer3d_game_data_world_trace_desc *trace,
                                      slayer3d_game_data_editor_selection *out_selection);
 bool editor_selection_mode_is_click(yyjson_val *selection);
+bool editor_mouse_in_rect(float mouse_x, float mouse_y, yyjson_val *rect);
+bool editor_handle_grid_widget(slayer3d_game_data_runtime *runtime, yyjson_val *editor, bool *out_consumed);
 void clear_editor_command_preview(slayer3d_game_data_runtime *runtime);
 void clear_editor_placement_preview(slayer3d_game_data_runtime *runtime);
 void update_editor_placement_preview(slayer3d_game_data_runtime *runtime, yyjson_val *editor,
